oko_sdk_react: simplify oko-sdk-react hooks API
