+++ sdk/oko_sdk_react/src/index.ts
@@ -1,21 +1,15 @@
-export { useAuth } from "./hooks/use_auth";
 export { useOko } from "./hooks/use_oko";
-export { useOkoModal } from "./hooks/use_oko_modal";
-export { useWalletInfo } from "./hooks/use_wallet_info";
 export { OkoProvider } from "./provider";
 export type {
   AuthType,
   OkoCoreState,
   OkoCosmosConfig,
   OkoEthConfig,
-  OkoInitError,
   OkoProviderConfig,
   OkoSvmConfig,
   OkoWalletInterface,
   OkoWalletState,
   SignInType,
-  UseAuthReturn,
-  UseOkoModalReturn,
   UseOkoReturn,
-  UseWalletInfoReturn,
+  WalletInfo,
 } from "./types";
